Fix GitHub PR warnings: circular import and missing cleanup logic
## Issues Fixed

### 🔄 Circular Import Dependency (Critical)
**Problem**: Circular import chain between pages/query.py and query/helpers/data_formatters.py:
- pages/query.py → query.callbacks → query.helpers.data_formatters → pages.query
- This would cause ImportError during module loading

**Solution**:
- Moved convert_phenotypic_to_behavioral_filters() from pages/query.py to query/helpers/data_formatters.py
- Updated pages/query.py to import from helper module instead of defining locally
- Verified no imports from pages module exist in query/ package

### 🧹 Missing Column Store Cleanup Logic (Data Integrity)
**Problem**: update_selected_columns_store callback missing cleanup for deselected tables:
- When tables removed from table-multiselect, their column selections persisted
- Led to incorrect export behavior with stale column selections

**Solution**:
- Added table-multiselect value as State to update_selected_columns_store callback
- Implemented cleanup logic to remove entries for deselected tables
- Added logging for cleanup operations for debugging

## Technical Changes
- **query/helpers/data_formatters.py**: Added complete convert_phenotypic_to_behavioral_filters function
- **pages/query.py**: Replaced function definition with import from helper module
- **query/callbacks/data_loading.py**: Enhanced update_selected_columns_store with cleanup logic

## Verification
- ✅ No circular imports: Tested successful import of helper function
- ✅ Clean import chain: query/ package has no imports from pages module
- ✅ Proper cleanup: Column selections removed when tables deselected
- ✅ Maintains backward compatibility: All existing functionality preserved
